Add support for faking gql by query name
