+++ .vscode/settings.json
@@ -23,5 +23,6 @@
     "eslint.provideLintTask": true,
     "[json]": {
         "editor.defaultFormatter": "esbenp.prettier-vscode"
-    }
+    },
+    "typescript.tsdk": "node_modules/typescript/lib"
 }
