+++ package.json
@@ -44,6 +44,9 @@
         "sharp": "^0.22.1",
         "uuid": "^3.3.2"
     },
+    "peerDependencies": {
+        "graphql": "*"
+    },
     "keywords": [
         "gatsby"
     ],
@@ -83,7 +86,7 @@
         "@testing-library/react-hooks": "^2.0.1",
         "@types/classnames": "^2.2.7",
         "@types/jest": "^23.3.12",
-        "@types/node": "^12.7.3",
+        "@types/node": "^12.12.14",
         "@types/react-dom": "^16.0.11",
         "@types/react-helmet": "^5.0.8",
         "@types/react-test-renderer": "^16.0.3",
@@ -100,7 +103,7 @@
         "babel-preset-gatsby": "^0.2.12",
         "codecov": "^3.5.0",
         "css-loader": "^0.28.11",
-        "eslint": "^6.6.0",
+        "eslint": "^6.7.1",
         "eslint-plugin-import": "^2.18.0",
         "eslint-plugin-prettier": "^3.1.0",
         "extract-text-webpack-plugin": "^3.0.2",
@@ -121,7 +124,7 @@
         "regenerator-runtime": "^0.12.1",
         "ts-jest": "^24.0.2",
         "ts-loader": "4",
-        "typescript": "^3.2.2",
+        "typescript": "^3.7.2",
         "typings-for-css-modules-loader": "^1.7.0"
     },
     "lint-staged": {
